qgs3daxis: Do not update the objects if the scene has not been created
When a renderview is created it handle the window resizes by calling
`updateWindowResize()`. In the case of the 3d axis, the renderview
calls `onViewportSizeUpdate()` which itself calls
`Qgs3DAxis::onAxisViewportSizeUpdate()`. This functions updates the
the cube and axis visibility. However, when the renderview is created
the scene has not been created yet and the cube and axis do not exist
yet. Therefore, this results in a segmentation fault.

This issue is fixed by not doing anything in
`Qgs3DAxis::onAxisViewportSizeUpdate()` if the 3d axis scene has not
been created yet.

Fixes: 75ee85e2309b5aa87e358355c22cf8fdafb87411